Functionality updates

+++ Application/Features/Book/Dtos/SimilarBookDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace Application.Features.Book.Dtos;
 
-public record SimilarBookDto(int BookId, string Title, string Author, double AverageRating, string CoverImageUrl, bool IsNewBook);
+public record SimilarBookDto(int BookId, string Title, string Author, decimal AverageRating, string CoverImageUrl, bool IsNewBook);
